Revert "Revert "Merge branch 'Developer' into MarcusNew""
This reverts commit 877be012c3968afe46f98b9777cc7024663365e5.

+++ PetShopAPI/Controllers/BrandController.cs
@@ -1,6 +1,8 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
 using PetShop.Domain;
 using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PetShop.Application.Interfaces;
 using PetShop.Application.PostProdDTO;
@@ -41,7 +43,7 @@ namespace PetShopApi.Controllers{
         {
             try
             {
-                return Ok(_brandService.GetBrandByID(brandID));
+                return Ok(_brandService.GetBrandById(brandID));
             }
             catch (KeyNotFoundException e)
             {
@@ -54,12 +56,21 @@ namespace PetShopApi.Controllers{
         }
 
         [HttpPost]
+        [Authorize]
         public ActionResult<Brand> CreateBrand(BrandDto dto)
         {
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                var result = _brandService.CreateBrand(dto);
-                return Created("shop/" + result.Id, result);
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    var result = _brandService.CreateBrand(dto);
+                    return Created("shop/" + result.Id, result);
+                }
             }
             catch (ValidationException e)
             {
@@ -73,12 +84,20 @@ namespace PetShopApi.Controllers{
 
         [HttpPut]
         [Route("{brandID}")]
-
+        [Authorize]
         public ActionResult<MainCategory> UpdateBrand([FromRoute] int brandID, [FromBody] BrandDto brandDto)
         {
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                return Ok(_brandService.UpdateBrand(brandID , brandDto));
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    return Ok(_brandService.UpdateBrand(brandID, brandDto));
+                }
             }
             catch (KeyNotFoundException e)
             {
@@ -91,12 +110,20 @@ namespace PetShopApi.Controllers{
         }
 
         [HttpDelete("{brandID}")]
-
+        [Authorize]
         public ActionResult<Brand> DeleteBrandByID(int brandID)
         {
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                return Ok(_brandService.DeleteBrand(brandID));
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    return Ok(_brandService.DeleteBrand(brandID));
+                }
             }
             catch (KeyNotFoundException e)
             {
