+++ PetShopAPI/Controllers/MainCatController.cs
@@ -1,5 +1,7 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
 using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PetShop.Application;
 using PetShop.Application.Interfaces;
@@ -60,12 +62,21 @@ namespace PetShopApi.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public ActionResult<MainCategory> CreateMainCategory(MainCatDTO dto)
         {
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                var result = _catService.CreateMainCategory(dto);
-                return Created("shop/" + result.MainCategoryID, result);
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    var result = _catService.CreateMainCategory(dto);
+                    return Created("shop/" + result.MainCategoryID, result);
+                }
             }
             catch (ValidationException e)
             {
@@ -79,13 +90,22 @@ namespace PetShopApi.Controllers
 
         [HttpPut]
         [Route("{mainCatID}")]
+        [Authorize]
 
         public ActionResult<MainCategory> UpdateMainCategory([FromRoute] int mainCatID, [FromBody] MainCatDTO mainCategory)
         {
             
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                return Ok(_catService.UpdateMainCategory(mainCatID,mainCategory));
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    return Ok(_catService.UpdateMainCategory(mainCatID, mainCategory));
+                }
             }
             catch (KeyNotFoundException e)
             {
@@ -98,12 +118,21 @@ namespace PetShopApi.Controllers
         }
 
         [HttpDelete("{mainCatID}")]
+        [Authorize]
 
         public ActionResult<MainCategory> DeleteMainCategoryByID(int mainCatID)
         {
+            // checking if the token holds an admin
+            bool hasClaim = User.HasClaim(ClaimTypes.Role, "Admin");
             try
             {
-                return Ok(_catService.DeleteMainCategoryById(mainCatID));
+                // Ensure the user is authenticated
+                if (!User.Identity.IsAuthenticated)
+                    return Unauthorized();
+                else
+                {
+                    return Ok(_catService.DeleteMainCategoryById(mainCatID));
+                }
             }
             catch (KeyNotFoundException e)
             {
